+++ src/app/services/cheatsheets.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CheatsheetsService } from './cheatsheets.service';
+
+describe('CheatsheetsService', () => {
+  let service: CheatsheetsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CheatsheetsService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});