+++ src/app/tools/base64/base64.component.ts
@@ -1,42 +1,52 @@
 import { Component, OnInit } from '@angular/core';
+import { FavoritesService } from 'src/app/services/favorites.service';
 
 @Component({
   selector: 'app-base64',
   templateUrl: './base64.component.html',
-  styleUrls: ['./base64.component.scss']
+  styleUrls: ['./base64.component.scss'],
 })
 export class Base64Component implements OnInit {
+  input: string = '';
+  output: string = '';
+  action: string = 'encode';
 
-  textToEncode: string = '';
-  textToDecode: string = '';
+  showOutput = false;
+  isFav = false;
 
-  showEncodedText = false;
-  showDecodedText = false;
-
-  encoded = '';
-  decoded = '';
-
-  constructor() { }
-
-  ngOnInit(): void {
+  constructor(private favS: FavoritesService) {
+    this.isFav = this.favS.isFavorite('base64');
   }
 
-  private toBase64(input: string){
+  ngOnInit(): void {}
+
+  private toBase64(input: string) {
     return btoa(input);
   }
 
-  private fromBase64(input: string){
+  private fromBase64(input: string) {
     return atob(input);
   }
 
-  encode(){
-    this.showEncodedText = true;
-    this.encoded = this.toBase64(this.textToEncode);
+  addToFav() {
+    let fav = {
+      name: 'Base64',
+      description: 'Encode and decode strings or images with Base64',
+      category: 'Encoding/Decoding',
+      icon: 'assets/icons/base64.svg',
+      url: 'base64',
+    };
+
+    this.isFav = true;
+    this.favS.addToFavorites(fav);
   }
 
-  decode(){
-    this.showDecodedText = true;
-    this.decoded = this.fromBase64(this.textToDecode);
+  handleInput() {
+    this.showOutput = true;
+    if (this.action.toLowerCase() == 'decode') {
+      this.output = this.fromBase64(this.input);
+    } else {
+      this.output = this.toBase64(this.input);
+    }
   }
-
 }