+++ CHANGELOG.md
@@ -5,10 +5,20 @@ All notable changes to this project will be documented in this file.
 
 ### Added
  - Bootstrap Libary
+ - Tools Overview Page
+ - Favorite Apps Page
+ - Tool Interface and Service
+ - Cheatsheets Overview Page
+ - Cheatsheet Interface and Service
+ - HTML Entitiew Cheatsheet
+ - CSS Cursor Cheatsheet
 
  ### Changed
  - Replaced Angular Material mat-toolbar with bootstrap nav
  - Replaced class 'card' with 'base64-card' to prevent bootstrap override
+
+ ## Removed
+ - home page (landing page is now the favorite page)
 
 ## [0.0.1]